refactor: 공통으로 사용되는 `PaymentCard` 컴포저블 및 `CardUiModel` 위치 변경
