+++ app/src/androidTest/java/woowacourse/payments/ui/cardwallet/PaymentCardTest.kt
@@ -7,8 +7,8 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.payments.ui.model.CardUiModel
-import woowacourse.payments.ui.newcard.components.PaymentCard
+import woowacourse.payments.ui.common.model.CardUiModel
+import woowacourse.payments.ui.common.model.components.PaymentCard
 
 class PaymentCardTest {
     @get:Rule
