+++ app/src/main/java/woowacourse/payments/ui/cardwallet/components/CardWalletContent.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import woowacourse.payments.designsystem.theme.AndroidpaymentsTheme
 import woowacourse.payments.ui.cardwallet.model.CardWalletState
-import woowacourse.payments.ui.model.CardUiModel
+import woowacourse.payments.ui.common.model.CardUiModel
 
 @Composable
 fun CardWalletContent(
